requirements.txt
cloudpickle==3.0.0
contourpy==1.3.0
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.53.1
imbalanced-learn==0.12.3
imblearn==0.0
joblib==1.4.2
kaleido==0.2.1
kiwisolver==1.4.5
llvmlite==0.43.0
matplotlib==3.9.2
mlxtend==0.23.1
numba==0.60.0
numpy==1.26.4
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.24.0
pynndescent==0.5.13
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
shap==0.46.0
six==1.16.0
slicer==0.0.8
tenacity==9.0.0
threadpoolctl==3.5.0
tqdm==4.66.5
tzdata==2024.1
umap-learn==0.5.6
